- Added extension for validating data annotations. - Modified the headered grid controls. - Modified controls for skipping the tab stop. - Modified  KarveRoutingBaseViewModel for allowing CanExecuteCommand. In this way when the UI validations fails it will not execute. - Added a validation chain inside the booking data saver. Chain of responsability design pattern. - Added the KarveBusinessServices Project for the business layer. - Added several validation rules: a concept shall be not null, a date shall be correct, a group shall be not null. Other fields shall be not null. - Added static class DateUtils for ULTIMODI - Added BookingUnitToIndexConverter for the booking unit.

+++ src/Modules/DALModule/KarveDataServices/DataTransferObject/BookingSummaryDto.cs
@@ -178,10 +178,10 @@ namespace KarveDataServices.DataTransferObject
         public string MultipleDirection { set; get; }
         [Display(Name = "Total deposito", Description = "Total")]
     
-        public ulong DepositTotal { set; get; }
+        public decimal DepositTotal { set; get; }
         [Display(Name = "Deposito", Description = "Deposito")]
 
-        public ulong Deposit { set; get; }
+        public decimal Deposit { set; get; }
         /// <summary>
         ///  Set or get the notes.
         /// </summary>
